Updated scripts, added a check for the existence of the docs folder

+++ rename-browser-to-docs.js
@@ -7,12 +7,23 @@ async function renameFolder() {
     const newPath = path.join(distPath, 'docs');
 
     try {
-        const exists = await fs.pathExists(oldPath);
-        console.log(`Does ${oldPath} exist? ${exists}`);
+        const existsOld = await fs.pathExists(oldPath);
+        const existsNew = await fs.pathExists(newPath);
+        
+        console.log(`Does ${oldPath} exist? ${existsOld}`);
+        console.log(`Does ${newPath} exist? ${existsNew}`);
 
-        await fs.remove(newPath);
-        await fs.rename(oldPath, newPath);
-        console.log(`Successfully renamed ${oldPath} to ${newPath}`);
+        if (existsNew) {
+            console.log(`Removing existing ${newPath}`);
+            await fs.remove(newPath);
+        }
+
+        if (existsOld) {
+            await fs.rename(oldPath, newPath);
+            console.log(`Successfully renamed ${oldPath} to ${newPath}`);
+        } else {
+            console.log(`${oldPath} does not exist, skipping rename.`);
+        }
     } catch (err) {
         console.error(err);
     }
